pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "topoprompt"
version = "0.1.0"
description = "Compiler for discovering minimal prompt-program topologies."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "TopoPrompt" }
]
dependencies = [
  "datasets>=4.0.0",
  "diskcache>=5.6.3",
  "networkx>=3.4",
  "numpy>=2.0.0",
  "openai>=2.0.0",
  "orjson>=3.10.0",
  "pandas>=2.2.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "scikit-learn>=1.6.0",
]

[project.optional-dependencies]
dspy = ["dspy>=2.6.0"]
dev = [
  "pytest>=8.4.2",
]

[project.scripts]
topoprompt = "topoprompt.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["topoprompt"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

